+++ src/app/hooks/useCommands.ts
@@ -29,6 +29,7 @@ import { getStateEvent } from '$utils/room';
 import { splitWithSpace } from '$utils/common';
 import { useSetting } from '$state/hooks/settings';
 import { settingsAtom } from '$state/settings';
+import { useOpenBugReportModal } from '$state/hooks/bugReportModal';
 import { createRoomEncryptionState } from '$components/create-room';
 import { parsePronounsInput } from '$utils/pronouns';
 import { useRoomNavigate } from './useRoomNavigate';
@@ -241,6 +242,15 @@ export enum Command {
   SetExt = 'setext',
   DelExt = 'delext',
   DiscardSession = 'discardsession',
+  // Cute Events
+  Hug = 'hug',
+  Cuddle = 'cuddle',
+  // our own cute events, not part of FluffyChat or other clients
+  Wave = 'wave',
+  Poke = 'poke',
+  Headpat = 'headpat',
+  // Meta
+  Report = 'bugreport',
 }
 
 export type CommandContent = {
@@ -255,6 +265,7 @@ export const useCommands = (mx: MatrixClient, room: Room): CommandRecord => {
   const { navigateRoom } = useRoomNavigate();
   const [developerTools] = useSetting(settingsAtom, 'developerTools');
   const profile = useUserProfile(mx.getSafeUserId());
+  const openBugReport = useOpenBugReportModal();
 
   const commands: CommandRecord = useMemo(
     () => ({
@@ -1295,8 +1306,93 @@ export const useCommands = (mx: MatrixClient, room: Room): CommandRecord => {
           }
         },
       },
+      // Cute Events
+      [Command.Hug]: {
+        name: Command.Hug,
+        description: 'Send a hug to someone. Example: /hug [@user:example.org]',
+        exe: async (payload) => {
+          const target = payload.trim();
+          await mx.sendMessage(room.roomId, {
+            msgtype: 'im.fluffychat.cute_event',
+            'm.mentions': {
+              user_ids: target ? [target] : [],
+            },
+            cute_type: 'hug',
+            body: `🤗`,
+          } as any);
+        },
+      },
+      [Command.Cuddle]: {
+        name: Command.Cuddle,
+        description: 'Send a cuddle to someone. Example: /cuddle [@user:example.org]',
+        exe: async (payload) => {
+          const target = payload.trim();
+          await mx.sendMessage(room.roomId, {
+            msgtype: 'im.fluffychat.cute_event',
+            cute_type: 'cuddle',
+            'm.mentions': {
+              user_ids: target ? [target] : [],
+            },
+            body: `😊`,
+          } as any);
+        },
+      },
+      [Command.Wave]: {
+        name: Command.Wave,
+        description: 'Send a wave to someone. Example: /wave [@user:example.org]',
+        exe: async (payload) => {
+          const target = payload.trim();
+          await mx.sendMessage(room.roomId, {
+            msgtype: 'im.fluffychat.cute_event',
+            cute_type: 'wave',
+            'm.mentions': {
+              user_ids: target ? [target] : [],
+            },
+            body: `👋`,
+          } as any);
+        },
+      },
+      [Command.Poke]: {
+        name: Command.Poke,
+        description: 'Send a poke to someone. Example: /poke [@user:example.org]',
+        exe: async (payload) => {
+          const target = payload.trim();
+          await mx.sendMessage(room.roomId, {
+            msgtype: 'im.fluffychat.cute_event',
+            cute_type: 'poke',
+            'm.mentions': {
+              user_ids: target ? [target] : [],
+            },
+            body: `🫵`,
+          } as any);
+        },
+      },
+      [Command.Headpat]: {
+        name: Command.Headpat,
+        description: 'Send a headpat to someone. Example: /headpat [@user:example.org]',
+        // not really like any of the other cute events, but it was too good not to include
+        // using a custom msgtype to avoid confusion with the other existing cute events
+        exe: async (payload) => {
+          const target = payload.trim();
+          await mx.sendMessage(room.roomId, {
+            msgtype: 'fyi.cisnt.headpat',
+            'm.mentions': {
+              user_ids: target ? [target] : [],
+            },
+            body: `*pat pat*`,
+          } as any);
+        },
+      },
+      // Meta commands
+      [Command.Report]: {
+        name: Command.Report,
+        description: 'Report a bug or request a feature',
+        exe: async () => {
+          openBugReport();
+        },
+      },
     }),
-    [mx, navigateRoom, room, profile.displayName, profile.avatarUrl, developerTools]
+    [mx, navigateRoom, room, profile.displayName, profile.avatarUrl, developerTools, openBugReport]
   );
 
   return commands;