+++ src/app/state/settings.ts
@@ -78,6 +78,7 @@ export interface Settings {
   captionPosition: CaptionPosition;
 
   // Sable features!
+  sendPresence: boolean;
   mobileGestures: boolean;
   rightSwipeAction: RightSwipeAction;
   hideMembershipInReadOnly: boolean;
@@ -153,11 +154,12 @@ const defaultSettings: Settings = {
   captionPosition: CaptionPosition.Below,
 
   // Sable features!
+  sendPresence: true,
   mobileGestures: true,
   rightSwipeAction: RightSwipeAction.Reply,
   hideMembershipInReadOnly: true,
   useRightBubbles: false,
-  showUnreadCounts: true,
+  showUnreadCounts: false,
   badgeCountDMsOnly: false,
   showPingCounts: true,
   hideReads: false,